Use the NVS for wifi network and details on functions
Wifi name and password is now stored in NVS
Functions (ie number of lights etc) and pin assignments now read from NVS.
The intention is to make it easier to upgrade software and only a single build is required for any function.

+++ components/doorbell/doorbell.c
@@ -12,7 +12,7 @@ static const char pressed[] = "pressed";
 static iotElement_t element;
 static iotElementPub_t alertPub;
 
-void doorbellInit(int8_t pin)
+void doorbellInit(int pin)
 {
     switchAdd(pin, doorbellSwitchCallback, NULL);
     element.name = "doorbell";
